feat(codebase)!: adding the codebase
Initial commit of the ParkingManager project to add
the basic structure of a .NET (C#) project.

This project is meant to explore C# programming through
a small project not to make a business ready solution.

As such, the code is not to be expected to be of high
quality nor should it be considered professional.